python/requirements.txt
paho-mqtt
pymodbus==2.3.0
graphyte
